changing definition of prediction
This is the latest code. If continue to develop, start here instead of "development" branch. Be sure to also check out code via HiPerGator.

Going forward, we'll not only use a different definition for prediction, we'll also define observations and possibly rules.

The original paper (with this code base) was for prediction recognition. Now, the goal is prediction analysis--how accurate are prediction.