link prediction, observation with score
match(p_a=profit, o_a=profit) => Get the sentences where we have the word "profit" in the prediction and "profit" in the observation. Then, get how semantically similar the sentences are. Decided to only apply the attribute match for now, so this is a partial match and not exact as if exact, we'll match the dates and target entities.